Fix bootstrap script for older AMI versions
Source of fix: https://forums.aws.amazon.com/message.jspa?messageID=614195

With the old mrjob configuration, the commands in the ReadMe don't run out of
the box - instance boostrapping crashes with a "transaction check error". The
issue arises due to changes in the yum package environment and its dependencies.
Specifying a particular yum version (see this commit) fixes the problem.